Fix cBioDataPack error on non-char Patient ID
cBioDataPack() smartly checks patient IDs for the prefix "TCGA" to remove it if
needed. That checks errors when the patient ID is not numeric.
Not all studieas' patient IDs are characters. For example,
coadread_genentech patient IDs are numeric. Modifying the test
to first check if the IDs are character solves the problem.