Split cBioDataPack func into download and parse
This will allow calling of just the parsing function instead of
downloading+parsing in one function. This is part of the effort to run
testing as a github action on the datahub repo directly. On datahub we
won't need to download the study, only need to see if we can parse it